Revert "feat(pipelined): enable NAT traffic for eNB traffic in Non-NAT setup (#8134)"
This reverts commit 0d43b150c5f3674bc60712df593149fefca47c0a.
