+++ lte/gateway/python/magma/pipelined/main.py
@@ -21,7 +21,7 @@ import threading
 
 import aioeventlet
 from lte.protos.mconfig import mconfigs_pb2
-from magma.common.misc_utils import get_ip_from_if
+from magma.common.misc_utils import call_process, get_ip_from_if
 from magma.common.sentry import sentry_init
 from magma.common.service import MagmaService
 from magma.configuration import environment
@@ -29,7 +29,7 @@ from magma.pipelined.app import of_rest_server
 from magma.pipelined.app.he import PROXY_PORT_NAME
 from magma.pipelined.bridge_util import BridgeTools
 from magma.pipelined.check_quota_server import run_flask
-from magma.pipelined.datapath_setup import tune_datapath, setup_masquerade_rule
+from magma.pipelined.datapath_setup import tune_datapath
 from magma.pipelined.gtp_stats_collector import (
     MIN_OVSDB_DUMP_POLLING_INTERVAL,
     GTPStatsCollector,
@@ -121,7 +121,6 @@ def main():
 
     # tune datapath according to config
     tune_datapath(service.config)
-    setup_masquerade_rule(service.config, service.loop)
 
     # monitoring related configuration
     mtr_interface = service.config.get('mtr_interface', None)
@@ -133,6 +132,23 @@ def main():
     service_manager = ServiceManager(service)
     service_manager.load()
 
+    def callback(returncode):
+        if returncode != 0:
+            logging.error(
+                "Failed to set MASQUERADE: %d", returncode,
+            )
+
+    # TODO fix this hack for XWF
+    if enable_nat is True or service.config.get('setup_type') == 'XWF':
+        ip_table_rule = 'POSTROUTING -o %s -j MASQUERADE' % service.config['nat_iface']
+        check_and_add = 'iptables -t nat -C %s || iptables -t nat -A %s' % \
+                (ip_table_rule, ip_table_rule)
+        logging.debug("check_and_add: %s", check_and_add)
+        call_process(
+            check_and_add,
+                callback,
+                service.loop,
+        )
 
     service.loop.create_task(
         monitor_ifaces(
